Closes #119; Localize input formatting
